# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Easypanel OAuth2 Server",
    "version": "1.0.0",
    "description": "Standards-compliant OAuth2 / OpenID Connect authorization server integrated with Easypanel user management.\n\n## Authentication\nMost OAuth endpoints use client credentials (Basic auth or body params).\nThe `/oauth/userinfo` endpoint requires a Bearer access token.\nThe `/api/clients` endpoints require either an authenticated admin session or an `X-API-Key` header.\n"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/oauth/authorize": {
      "get": {
        "tags": [
          "OAuth2"
        ],
        "summary": "Authorization endpoint",
        "description": "Initiates the Authorization Code flow. Redirects the user to the login page if not authenticated,\nthen to the consent screen. After approval, redirects back to `redirect_uri` with a `code` parameter.\n",
        "operationId": "authorize",
        "parameters": [
          {
            "name": "response_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "code"
              ]
            }
          },
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "openid profile email"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nonce",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to login or consent page"
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/oauth/userinfo": {
      "get": {
        "tags": [
          "OAuth2"
        ],
        "summary": "UserInfo endpoint (OIDC)",
        "description": "Returns claims about the authenticated user. Requires a valid Bearer access token.",
        "operationId": "userinfo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User claims",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/OAuthError"
          },
          "404": {
            "$ref": "#/components/responses/OAuthError"
          }
        }
      }
    },
...
    "/api/clients": {
      "get": {
        "tags": [
          "Client Management"
        ],
        "summary": "List all clients",
        "operationId": "listClients",
        "security": [
          {
            "apiKey": []
          },
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Client list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "clients": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Client"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      },
      "post": {
        "tags": [
          "Client Management"
        ],
        "summary": "Create a client",
        "operationId": "createClient",
        "security": [
          {
            "apiKey": []
          },
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client created (includes secret — only shown once)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "client": {
                      "$ref": "#/components/schemas/ClientWithSecret"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/clients/{id}": {
      "get": {
        "tags": [
          "Client Management"
        ],
        "summary": "Get a client",
        "operationId": "getClient",
        "security": [
          {
            "apiKey": []
          },
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientId"
          }
        ],
        "responses": {
          "200": {
            "description": "Client details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "client": {
                      "$ref": "#/components/schemas/Client"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      },
      "put": {
        "tags": [
          "Client Management"
        ],
        "summary": "Update a client",
        "operationId": "updateClient",
        "security": [
          {
            "apiKey": []
          },
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated client",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "client": {
                      "$ref": "#/components/schemas/Client"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      },
      "delete": {
        "tags": [
          "Client Management"
        ],
        "summary": "Delete a client",
        "operationId": "deleteClient",
        "security": [
          {
            "apiKey": []
          },
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientId"
          }
        ],
        "responses": {
          "200": {
            "description": "Client deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Client ID and secret as HTTP Basic credentials"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access token issued by the token endpoint"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Admin API key (CLIENT_API_KEY env var)"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie from browser login"
      }
    },
    "parameters": {
      "ClientId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "OAuth client ID"
      }
    },
    "responses": {
      "OAuthError": {
        "description": "OAuth error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/OAuthError"
            }
          }
        }
      },
      "ApiError": {
        "description": "API error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "schemas": {
...
      "UserInfo": {
        "type": "object",
        "properties": {
          "sub": {
            "type": "string",
            "description": "Subject identifier (user ID)"
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "redirectUris": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedScopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "persistent": {
            "type": "boolean"
          }
        }
      },
      "ClientWithSecret": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Client"
          },
          {
            "type": "object",
            "properties": {
              "secret": {
                "type": "string",
                "description": "Client secret — only returned on creation or secret regeneration"
              }
            }
          }
        ]
      },
      "ClientCreateRequest": {
        "type": "object",
        "required": [
          "name",
          "redirectUris",
          "allowedScopes"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "redirectUris": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "allowedScopes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "profile",
              "email",
              "openid"
            ]
          },
          "persistent": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "ClientUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "redirectUris": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedScopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "persistent": {
            "type": "boolean"
          }
        }
      },
...
      "OAuthError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "error_description": {
            "type": "string"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}